refactor: extract shared claim detail composables to ui-claim-status
